Improve queries

+++ types/response.go
@@ -34,7 +34,7 @@ type AccountGroupResponse struct {
 }
 type AccountPostResponse struct {
     Account                string `json:"account"`
-    TotalEffectiveNumUnits int64  `json:"totalEffectiveNumUnits"`
+    TotalEffectiveNumUnits uint32  `json:"totalEffectiveNumUnits"`
 }
 
 type Account struct {
